Ajout de logo et signature comme attribut d'etablissement et de medecin controle

+++ app/Http/Controllers/Api/ResultatImagerieController.php
@@ -139,6 +139,7 @@ class ResultatImagerieController extends Controller
             $this->uploadFile($request,$resultat);
         }
 
+        if (!is_null($file))
         File::delete(public_path().'/storage/'.$file);
 
         return response()->json([
